fix copyrights; extract `notebookExecutionQueue.ts`
