+++ src/vs/workbench/contrib/runtimeNotebookKernel/browser/runtimeNotebookKernel.ts
@@ -1,5 +1,5 @@
 /*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2024 Posit Software, PBC. All rights reserved.
+ *  Copyright (C) 2025 Posit Software, PBC. All rights reserved.
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
@@ -22,38 +22,9 @@ import { NotebookCellTextModel } from '../../notebook/common/model/notebookCellT
 import { INotebookExecutionStateService } from '../../notebook/common/notebookExecutionStateService.js';
 import { INotebookKernel, INotebookKernelChangeEvent, VariablesResult } from '../../notebook/common/notebookKernelService.js';
 import { INotebookService } from '../../notebook/common/notebookService.js';
+import { NotebookExecutionQueue } from '../common/notebookExecutionQueue.js';
 import { POSITRON_RUNTIME_NOTEBOOK_KERNELS_EXTENSION_ID } from '../common/runtimeNotebookKernelConfig.js';
 import { RuntimeNotebookCellExecution } from './runtimeNotebookCellExecution.js';
-
-/**
- * A queue for notebook executions, like SequencerByKey but keyed by URIs and stops the queue when
- * an execution errors.
- */
-class NotebookExecutionSequencer {
-	/** Map of last queued execution promise, keyed by notebook URI. */
-	private readonly _promisesByNotebookUri = new ResourceMap<Promise<unknown>>();
-
-	queue<T>(key: URI, promiseTask: () => Promise<T>): Promise<T> {
-		// Get the last queued promise for this notebook, if one exists.
-		const lastPromise = this._promisesByNotebookUri.get(key) ?? Promise.resolve();
-
-		// Chain the new promise after the last.
-		const newPromise = lastPromise
-			.then(promiseTask)
-			.finally(() => {
-				// If the last promise in the chain ended, delete the entry for the notebook,
-				// starting a new chain.
-				if (this._promisesByNotebookUri.get(key) === newPromise) {
-					this._promisesByNotebookUri.delete(key);
-				}
-			});
-
-		// Update the last promise for this notebook.
-		this._promisesByNotebookUri.set(key, newPromise);
-
-		return newPromise;
-	}
-}
 
 /** A notebook kernel for Positron's language runtimes. */
 export class RuntimeNotebookKernel extends Disposable implements INotebookKernel {
@@ -114,7 +85,7 @@ export class RuntimeNotebookKernel extends Disposable implements INotebookKernel
 	 * Each queued cell execution promise is chained to the previous one for the notebook
 	 * so that cells are executed sequentially.
 	 */
-	private readonly _notebookExecutionSequencer = new NotebookExecutionSequencer();
+	private readonly _notebookExecutionSequencer = new NotebookExecutionQueue();
 
 	/**
 	 * The current pending execution, keyed by notebook URI.
